Restructure chat-ui to look better for other devs
