+++ Frontend/chat-ui/src/components/inputbox/ChatInputBox.jsx
@@ -1,6 +1,3 @@
-import React from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-
 import {
     Button,
     Flex,
@@ -11,6 +8,8 @@ import {
     Box,
 } from '@chakra-ui/react';
 import { ArrowUpIcon } from '@chakra-ui/icons';
+import React from 'react';
+import { useSelector, useDispatch } from 'react-redux';
 
 import './ChatInputBox.css';
